Example Developer task overview
We have found that HCDP Development tasks work best when one developer is assigned to complete the task itself, and one developer takes on a task oversight role to provide feedback, define deliverables, and oversee the task to its completion. Generally, this type of discussion between the developer completing the task and the developer overseeing its completion is conducted over email or [Discord](https://discordapp.com/invite/gxAUagw). The attached document is an email thread between Noah and a developer completing an HCDP task to provide an example of how this communication occurs in practice.